Dynamic menu loading system - limited features
Established a dynamic menu loading system with limited capabilities.

Here, the menu items will be stored in the database, and based on permissions given to them, they will be filtered and returned as JSON objects to the frontend.

* A "username" column was added to the "permission" table to temporarily bypass the authorization complexities and focus on the menu filtering process instead. It will be removed in the next commit.

There are several places of improvement including better entity designing, standard permission procedure application etc.